fix incorrect info in man pages
If you tried reading the man pages, it told you that the manpages are
NOT the correct way to read the documentation, and recommended using
installing emacs and then using the emacs-specific "info" pages, or
alternatively the half-baked "info" command that doesn't really work
properly.

But, there are no texinfo pages for pfl which makes the whole thing
quite weird.